Fix some accented names
Anduril has been in here a while as has Barad-Dur but I see both being rejected
in the last 30 days. I think because what gets spat out is three crazy chars and
not two so I altered the regexes to account for that.

We should probably come up with some general solve for this but this is all the
problems I saw from the last 30 days after spot-checking a bunch.